moved solr client creation to a factory
